feat(crd generation): Add additional printer columns. (#118)
This closes #60. An additional `[AdditionalPrinterColumn]`
adds the possibility to add a specific property as an
additional printer column. The additional
`[GenericAdditionalPrinterColumn]` adds the possibility
to add any spec path to the additional printer columns.

Additionally, documentation for the various
property and class attributes for entities was
added.

+++ src/KubeOps/Operator/Entities/Annotations/LengthAttribute.cs
@@ -2,11 +2,22 @@
 
 namespace KubeOps.Operator.Entities.Annotations
 {
+    /// <summary>
+    /// Defines length limits for properties.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Property)]
     public class LengthAttribute : Attribute
     {
-        public int MinLength { get; set; } = -1;
+        /// <summary>
+        /// If not `-1`, define the minimum length.
+        /// -1 is used as the "null" - value.
+        /// </summary>
+        public long MinLength { get; set; } = -1;
 
-        public int MaxLength { get; set; } = -1;
+        /// <summary>
+        /// If not `-1`, define the maximum length.
+        /// -1 is used as the "null" - value.
+        /// </summary>
+        public long MaxLength { get; set; } = -1;
     }
 }
